Add files via upload
Write a C++ menu-driven program that calculates and displays the area of a square, cube, rectangle, and cuboid. Consider length as the side value for the square and cuboid. Identify proper data members and member functions. Design and create an appropriate class for the given scenario.  Maintain proper boundary conditions and follow coding best practices. The menus are as follows,
 a) Square
 b) Cube
 c) Rectangle
 d) Cuboid
 e) Exit