Nene's Game
Nene invented a new game based on an increasing sequence of integers 𝑎1,𝑎2,…,𝑎𝑘
.

In this game, initially 𝑛
 players are lined up in a row. In each of the rounds of this game, the following happens:

Nene finds the 𝑎1
-th, 𝑎2
-th, …
, 𝑎𝑘
-th players in a row. They are kicked out of the game simultaneously. If the 𝑖
-th player in a row should be kicked out, but there are fewer than 𝑖
 players in a row, they are skipped.
Once no one is kicked out of the game in some round, all the players that are still in the game are declared as winners.

For example, consider the game with 𝑎=[3,5]
 and 𝑛=5
 players. Let the players be named player A, player B, …
, player E in the order they are lined up initially. Then,

Before the first round, players are lined up as ABCDE. Nene finds the 3
-rd and the 5
-th players in a row. These are players C and E. They are kicked out in the first round.
Now players are lined up as ABD. Nene finds the 3
-rd and the 5
-th players in a row. The 3
-rd player is player D and there is no 5
-th player in a row. Thus, only player D is kicked out in the second round.
In the third round, no one is kicked out of the game, so the game ends after this round.
Players A and B are declared as the winners.
Nene has not yet decided how many people would join the game initially. Nene gave you 𝑞
 integers 𝑛1,𝑛2,…,𝑛𝑞
 and you should answer the following question for each 1≤𝑖≤𝑞
 independently:

How many people would be declared as winners if there are 𝑛𝑖
 players in the game initially?
Input
Each test contains multiple test cases. The first line contains the number of test cases 𝑡
 (1≤𝑡≤250
). The description of test cases follows.

The first line case contains two integers 𝑘
 and 𝑞
 (1≤𝑘,𝑞≤100
) — the length of the sequence 𝑎
 and the number of values 𝑛𝑖
 you should solve this problem for.

The second line contains 𝑘
 integers 𝑎1,𝑎2,…,𝑎𝑘
 (1≤𝑎1<𝑎2<…<𝑎𝑘≤100
) — the sequence 𝑎
.

The third line contains 𝑞
 integers 𝑛1,𝑛2,…,𝑛𝑞
 (1≤𝑛𝑖≤100
).

Output
For each test case, output 𝑞
 integers: the 𝑖
-th (1≤𝑖≤𝑞
) of them should be the number of players declared as winners if initially 𝑛𝑖
 players join the game.