Advantage
There are 𝑛
 participants in a competition, participant 𝑖
 having a strength of 𝑠𝑖
.

Every participant wonders how much of an advantage they have over the other best participant. In other words, each participant 𝑖
 wants to know the difference between 𝑠𝑖
 and 𝑠𝑗
, where 𝑗
 is the strongest participant in the competition, not counting 𝑖
 (a difference can be negative).

So, they ask you for your help! For each 𝑖
 (1≤𝑖≤𝑛
) output the difference between 𝑠𝑖
 and the maximum strength of any participant other than participant 𝑖
.

Input
The input consists of multiple test cases. The first line contains an integer 𝑡
 (1≤𝑡≤1000
) — the number of test cases. The descriptions of the test cases follow.

The first line of each test case contains an integer 𝑛
 (2≤𝑛≤2⋅105
) — the length of the array.

The following line contains 𝑛
 space-separated positive integers 𝑠1
, 𝑠2
, ..., 𝑠𝑛
 (1≤𝑠𝑖≤109
) — the strengths of the participants.

It is guaranteed that the sum of 𝑛
 over all test cases does not exceed 2⋅105
.

Output
For each test case, output 𝑛
 space-separated integers. For each 𝑖
 (1≤𝑖≤𝑛
) output the difference between 𝑠𝑖
 and the maximum strength of any other participant.