Remove prefix
Polycarp was presented with some sequence of integers 𝑎
 of length 𝑛
 (1≤𝑎𝑖≤𝑛
). A sequence can make Polycarp happy only if it consists of different numbers (i.e. distinct numbers).

In order to make his sequence like this, Polycarp is going to make some (possibly zero) number of moves.

In one move, he can:

remove the first (leftmost) element of the sequence.
For example, in one move, the sequence [3,1,4,3]
 will produce the sequence [1,4,3]
, which consists of different numbers.

Determine the minimum number of moves he needs to make so that in the remaining sequence all elements are different. In other words, find the length of the smallest prefix of the given sequence 𝑎
, after removing which all values in the sequence will be unique.

Input
The first line of the input contains a single integer 𝑡
 (1≤𝑡≤104
) — the number of test cases.

Each test case consists of two lines.

The first line contains an integer 𝑛
 (1≤𝑛≤2⋅105
) — the length of the given sequence 𝑎
.

The second line contains 𝑛
 integers 𝑎1,𝑎2,…,𝑎𝑛
 (1≤𝑎𝑖≤𝑛
) — elements of the given sequence 𝑎
.

It is guaranteed that the sum of 𝑛
 values over all test cases does not exceed 2⋅105
.

Output
For each test case print your answer on a separate line — the minimum number of elements that must be removed from the beginning of the sequence so that all remaining elements are different.